First Commit
I did everything myself up until this point. Had trouble with an error message when I tried pushing commits. I made a new repo to avoid this error and to make a fresh repo for my team. Will be working with two others to try to make a game for one of my courses