Fun Stuff
Worked on fixing the health system. Added clamps to stop the armor from going over 100 and the health from going in the negatives. Added a death function that gives the player a ragdoll death but currently falls through the floor. Added game over screen that is added to viewport when the player dies. Still working on that though