Hitmarker, Player
Edited the position of the ADS Hitmarker to make it more aligned. Added some branches for when the player dies in order to restrict the player from firing, reloading, crouching, etc when the player dies. Going home for Christmas for the next week but when I get back I plan to work on this game the rest of the break