Main Menu
Working hard on getting the main menu to look nice. Got the end screen quit button to work properly. Just need to keep adding to it. Added code to pause the player and show the cursor once you died.  Clicking quit takes you to a new level which is the main menu