Settings/Backstory
Worked on adding a backstory section to my main menu. I changed it to backstory last minute as I couldn't figure how to change, apply and save settings without having to spend some money. So I changed settings to backstory. In the blueprints, everything labeled with settings is for the backstory now.