Health System Works
I finally got the health to work how I wanted. Once you are out of combat (when you stop colliding with a zombie) then a timer will start. If the timer completes then just hit will be set to false and then the player will start regaining armor. If the timer has started but then you collide with a zombie then the timer will reset.