Collisions, Jump, Grenade
Added colliders and physics to some objects that needed it in the level. Increased the games gravity to make the player not jump as high and to come down quicker. Added a radial force to the grenades so they can throw some stuff around. Working on making it simulate physics for the zombie if they overlap with the grenade's sphere collider