Grenade Beauty
Made my grenades operate much much better. Added several for each loops that check for the player and zombies. First for each will kill the player if he overlaps with the grenade sphere collider. Next one simulates the physics for the zombies if they overlap with the grenade collider and the last one kills the zombie if it overlaps. Had to do two for the zombie checker because I dont have a death function for them and I used the overlap component to sim the physics and the overlap actors to take the life variable away from the zombie