Sound, Grenade, Hitmarker
Mostly added some sounds to this commit. Gunshot, dryfire and reloading sounds. Want to change the drop magazine sound but for now its okay. Tweaked some of the numbers on the grenade and made it where if the zombie is dead and you shoot him then a hitmarker will not appear