Audio, Raytracing fix
Added audio for sprinting and getting hit. Made a timer for when the audio for panting should execute and end depending on when the player presses and releases the sprint key. Also made a sprint time out timer that doesn't allow the player to run for 10 seconds if they run for too long. Fixed the camera and gun so that the aiming down sights raytracing is much more accurate. This made the ak look a little weird but its okay.